+++ src/theme/stories/index.stories.js
@@ -4,7 +4,7 @@ import Box from 'ui-box'
 import Component from '@reactions/component'
 import { Heading } from '../../typography'
 import { Checkbox } from '../../checkbox'
-import v5Theme from '../../themes/v5'
+import v5Theme from '../../themes/classic'
 import v6Theme from '../../themes/default'
 import { TextInput } from '../../text-input'
 import { Textarea } from '../../textarea'
