+++ src/themes/classic/component-specific/checkbox.js
