+++ src/themes/classic/component-specific/index.js
