+++ src/themes/classic/component-specific/segmentedControl.js
