+++ src/themes/classic/index.js
