+++ src/themes/classic/typography/headings.js
