+++ src/themes/classic/typography/index.js
