+++ src/themes/classic/typography/paragraph.js
