+++ src/themes/index.js
@@ -1,2 +1,2 @@
 export { default as defaultTheme } from './default'
-export { default as classicTheme } from './v5'
+export { default as classicTheme } from './classic'
